// src/common/enums/condition.enum.ts
export enum Condition {
  GOOD = 'Good',
  FAIR = 'Fair',
  POOR = 'Poor',
}

export enum AssessmentType {
  INITIAL = 'Initial',
  PERIODIC = 'Periodic',
}

export enum AmenityValue {
  HIGH = 'High',
  MEDIUM = 'Medium',
  LOW = 'Low',
}

export enum TransplantSurvival {
  HIGH = 'High',
  MEDIUM = 'Medium',
  LOW = 'Low',
}